+++ src/fr/maxlego08/menu/zcore/utils/meta/Meta.java
@@ -2,6 +2,7 @@ package fr.maxlego08.menu.zcore.utils.meta;
 
 import fr.maxlego08.menu.api.utils.MetaUpdater;
 import fr.maxlego08.menu.save.Config;
+import fr.maxlego08.menu.zcore.logger.Logger;
 import fr.maxlego08.menu.zcore.utils.nms.NMSUtils;
 
 public class Meta {
@@ -16,8 +17,10 @@ public class Meta {
             try {
                 Class.forName("net.kyori.adventure.text.minimessage.MiniMessage");
                 meta = new ComponentMeta();
+                Logger.info("Use ComponentMeta");
             } catch (Exception e) {
                 meta = new ClassicMeta();
+                Logger.info("Use ClassicMeta");
             }
         }